chore: Move integration tests to separate source root
