+++ annotation-processor/src/it/java/com/bloxbean/cardano/client/plutus/annotation/it/ComplexTypesBlueprint.java
@@ -1,0 +1,7 @@
+package com.bloxbean.cardano.client.plutus.annotation.it;
+
+import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
+
+@Blueprint(fileInResources = "blueprint/ComplexTypeBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.complextypes")
+public interface ComplexTypesBlueprint {
+}
