+++ annotation-processor/src/it/java/com/bloxbean/cardano/client/plutus/annotation/it/HelloWorldBlueprint.java
@@ -1,11 +1,11 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
+package com.bloxbean.cardano.client.plutus.annotation.it;
 
 import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
 import com.bloxbean.cardano.client.plutus.annotation.ExtendWith;
 import com.bloxbean.cardano.client.quicktx.blueprint.extender.LockUnlockValidatorExtender;
 
-@Blueprint(fileInResources = "blueprint/multiple_validators.json", packageName = "com.test.multiple")
+@Blueprint(fileInResources = "blueprint/Hello_world-Blueprint.json", packageName = "com.test.hello")
 @ExtendWith(LockUnlockValidatorExtender.class)
-public interface MultipleValidatorsBlueprint {
+public interface HelloWorldBlueprint {
 
 }
