+++ annotation-processor/src/it/java/com/bloxbean/cardano/client/plutus/annotation/it/HelloWorldNoNSBlueprint.java
@@ -1,11 +1,10 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
+package com.bloxbean.cardano.client.plutus.annotation.it;
 
 import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
 import com.bloxbean.cardano.client.plutus.annotation.ExtendWith;
 import com.bloxbean.cardano.client.quicktx.blueprint.extender.LockUnlockValidatorExtender;
 
-@Blueprint(fileInResources = "blueprint/Hello_world-Blueprint.json", packageName = "com.test.hello")
+@Blueprint(fileInResources = "blueprint/HelloWorldNoNS.json", packageName = "com.test.hellons")
 @ExtendWith(LockUnlockValidatorExtender.class)
-public interface HelloWorldBlueprint {
-
+public interface HelloWorldNoNSBlueprint {
 }
