+++ annotation-processor/src/it/java/com/bloxbean/cardano/client/plutus/annotation/it/ListBlueprint.java
@@ -1,0 +1,7 @@
+package com.bloxbean.cardano.client.plutus.annotation.it;
+
+import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
+
+@Blueprint(fileInResources = "blueprint/ListBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.list")
+public interface ListBlueprint {
+}
