+++ annotation-processor/src/it/java/com/bloxbean/cardano/client/plutus/annotation/it/MapBlueprint.java
@@ -1,0 +1,7 @@
+package com.bloxbean.cardano.client.plutus.annotation.it;
+
+import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
+
+@Blueprint(fileInResources = "blueprint/MapBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.map")
+public interface MapBlueprint {
+}
