+++ annotation-processor/src/it/resources/blueprint/HelloWorldNoNS.json
@@ -1,0 +1,64 @@
+{
+  "$schema": "https://cips.cardano.org/cips/cip57/schemas/plutus-blueprint.json",
+
+  "$id": "https://github.com/aiken-lang/aiken/blob/main/examples/hello_world/plutus.json",
+
+  "$vocabulary": {
+    "https://json-schema.org/draft/2020-12/vocab/core": true,
+    "https://json-schema.org/draft/2020-12/vocab/applicator": true,
+    "https://json-schema.org/draft/2020-12/vocab/validation": true,
+    "https://cips.cardano.org/cips/cip57": true
+  },
+
+  "preamble": {
+    "title": "aiken-lang/hello_world",
+    "description": "Aiken contracts for project 'aiken-lang/hello_world'",
+    "version": "1.0.0",
+    "plutusVersion": "v2"
+  },
+
+  "validators": [
+    {
+      "title": "hello_world",
+      "datum": {
+        "title": "Datum",
+        "purpose": "spend",
+        "schema": {
+          "anyOf": [
+            {
+              "title": "Datum",
+              "dataType": "constructor",
+              "index": 0,
+              "fields": [
+                {
+                  "title": "owner",
+                  "dataType": "bytes"
+                }
+              ]
+            }
+          ]
+        }
+      },
+      "redeemer": {
+        "title": "Redeemer",
+        "schema": {
+          "anyOf": [
+            {
+              "title": "Redeemer",
+              "dataType": "constructor",
+              "index": 0,
+              "fields": [
+                {
+                  "title": "msg",
+                  "dataType": "bytes"
+                }
+              ]
+            }
+          ]
+        }
+      },
+      "compiledCode": "58ad0100003232322225333004323253330063372e646e64004dd7198009801002240009210d48656c6c6f2c20576f726c64210013233300100137586600460066600460060089000240206eb8cc008c00c019200022253335573e004294054ccc024cdc79bae300a00200114a226660060066016004002294088c8ccc0040052000003222333300a3370e008004016466600800866e0000d2002300d001001235573c6ea8004526165734ae855d11",
+      "hash": "5e1e8fa84f2b557ddc362329413caa3fd89a1be26bfd24be05ce0a02"
+    }
+  ]
+}
