+++ annotation-processor/src/it/resources/blueprint/MapBlueprint.json
@@ -1,0 +1,75 @@
+{
+  "preamble": {
+    "title": "kammerlo/mapBP",
+    "description": "Aiken contracts for project 'kammerlo/mapBP'",
+    "version": "0.0.0",
+    "plutusVersion": "v2",
+    "compiler": {
+      "name": "Aiken",
+      "version": "v1.0.21-alpha+4b04517"
+    },
+    "license": "Apache-2.0"
+  },
+  "validators": [
+    {
+      "title": "mapBP.mapBP",
+      "redeemer": {
+        "title": "datum",
+        "schema": {
+          "$ref": "#/definitions/mapBP~1Datum"
+        }
+      },
+      "compiledCode": "58a10100003232323232323232223253330064a229309b2b19299980319b87480000044c8c8c8c94ccc034c03c0084c926323300100100422533300f00114984c8cc00c00cc048008c8c94ccc038cdc3a400000226464a666026602a0042930b1b99375c602600260180042c601800260200022c6eb4c034004c034008dd6180580098020018b1802001118029baa001230033754002ae6955ceaab9e5573eae855d11",
+      "hash": "e62ab7064350e2980fa9a066208ff9576f1760e847513a0f737802ca"
+    }
+  ],
+  "definitions": {
+    "Int": {
+      "dataType": "integer"
+    },
+    "List$mapBP/NestedA": {
+      "dataType": "map",
+      "keys": {
+        "$ref": "#/definitions/mapBP~1NestedA"
+      },
+      "values": {
+        "$ref": "#/definitions/Int"
+      }
+    },
+    "String": {
+      "dataType": "#string"
+    },
+    "mapBP/Datum": {
+      "title": "Datum",
+      "anyOf": [
+        {
+          "title": "Datum",
+          "dataType": "constructor",
+          "index": 0,
+          "fields": [
+            {
+              "title": "map",
+              "$ref": "#/definitions/List$mapBP~1NestedA"
+            }
+          ]
+        }
+      ]
+    },
+    "mapBP/NestedA": {
+      "title": "NestedA",
+      "anyOf": [
+        {
+          "title": "NestedA",
+          "dataType": "constructor",
+          "index": 0,
+          "fields": [
+            {
+              "title": "msg",
+              "$ref": "#/definitions/String"
+            }
+          ]
+        }
+      ]
+    }
+  }
+}
