+++ annotation-processor/src/test/java/com/bloxbean/cardano/client/plutus/annotation/ConstrAnnotationProcessorTest.java
@@ -3,7 +3,6 @@ package com.bloxbean.cardano.client.plutus.annotation;
 import com.bloxbean.cardano.client.plutus.annotation.processor.ConstrAnnotationProcessor;
 import com.google.testing.compile.Compilation;
 import com.google.testing.compile.JavaFileObjects;
-import org.junit.Ignore;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
