--- annotation-processor/src/test/java/com/bloxbean/cardano/client/plutus/annotation/blueprint/model/AnyPlutusDataBlueprint.java
@@ -1,7 +1,0 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
-
-import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
-
-//@Blueprint(fileInResources = "blueprint/AnyPlutusDataBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.anyplutusdata")
-public interface AnyPlutusDataBlueprint {
-}
