--- annotation-processor/src/test/java/com/bloxbean/cardano/client/plutus/annotation/blueprint/model/ComplexTypesBlueprint.java
@@ -1,7 +1,0 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
-
-import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
-
-//@Blueprint(fileInResources = "blueprint/ComplexTypeBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.complextypes")
-public interface ComplexTypesBlueprint {
-}
