--- annotation-processor/src/test/java/com/bloxbean/cardano/client/plutus/annotation/blueprint/model/ListBlueprint.java
@@ -1,7 +1,0 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
-
-import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
-
-//@Blueprint(fileInResources = "blueprint/ListBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.list")
-public interface ListBlueprint {
-}
