--- annotation-processor/src/test/java/com/bloxbean/cardano/client/plutus/annotation/blueprint/model/MapBlueprint.java
@@ -1,7 +1,0 @@
-package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
-
-import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
-
-//@Blueprint(fileInResources = "blueprint/MapBlueprint.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.map")
-public interface MapBlueprint {
-}
