+++ annotation-processor/src/test/resources/blueprint/MultipleValidatorsBlueprint.java
@@ -1,7 +1,11 @@
 package com.bloxbean.cardano.client.plutus.annotation.blueprint.model;
 
 import com.bloxbean.cardano.client.plutus.annotation.Blueprint;
+import com.bloxbean.cardano.client.plutus.annotation.ExtendWith;
+import com.bloxbean.cardano.client.quicktx.blueprint.extender.LockUnlockValidatorExtender;
+import com.bloxbean.cardano.client.quicktx.blueprint.extender.MintValidatorExtender;
 
 @Blueprint(fileInResources = "blueprint/multiple_validators.json", packageName = "com.bloxbean.cardano.client.plutus.annotation.blueprint.multiple")
+@ExtendWith({LockUnlockValidatorExtender.class, MintValidatorExtender.class})
 public interface MultipleValidatorsBlueprint {
 }
